requirements.txt
fastapi==0.110.1
mysql-connector-python==8.3.0
python-dotenv==1.0.1
google-generativeai==0.4.1
uvicorn==0.29.0
bs4==0.0.2